math_induction seems to be working now that it has enough nonterminals to learn the correct grammar. sequence_distribution can now be used with the new grammar induction architecture.
To make sampling faster, I want to implement local sampling. It seems like this will be useful since the lack of nonterminals was apparently the underlying problem I had been running into. So this is a checkpoint commit before I begin editing the sampler/parser in parser.h to be able to perform local
sampling.